fix: Turned review pages into reusable components
