+++ modules/dashboard/component/reviews/reviewPages/BuyersView.tsx
@@ -1,13 +1,12 @@
 //* Removed some unnecessary imports
-
 import React, { useState, useEffect } from 'react';
-import NavDashBoard from '../../../../../modules/dashboard/component/Navbar';
+import NavDashBoard from '@modules/dashboard/component/Navbar';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import RatingCard from '@modules/dashboard/component/reviews/review-page/RatingCard';
 import RatingBar from '@modules/dashboard/component/reviews/review-page/RatingBar';
-import SellerReview from '@modules/dashboard/component/reviews/review-page/SellersReview';
+import Review from '../review-page/Review';
 import Filter from '@modules/dashboard/component/reviews/review-page/ReviewFilter';
 import Pagination from '@ui/Pagination';
 import MainLayout from '../../../../../components/Layout/MainLayout';
@@ -15,12 +14,11 @@ import EmptyReviewPage from '@modules/dashboard/component/reviews/review-page/Em
 import Container from '@modules/auth/component/Container/Container';
 import Loader from '@ui/Loader';
 import { ReviewData, ReviewApiResponse, RatsData } from '../../../../../@types';
-import CategoriesNav from '@modules/marketplace/component/CategoriesNav/CategoriesNav';
-import Review from '../review-page/Review';
-import useCategoryNav from '@modules/marketplace/hooks/useCategoryNav';
+import { set } from 'nprogress';
 
 //* Moved type definitions to @types/index.d.ts
 const BuyersView = () => {
+  //* Added a function to set the page number in the url
   const setPage = (pageNumber: number) => {
     setCurrentPage(pageNumber);
     router.push({
@@ -35,11 +33,11 @@ const BuyersView = () => {
   //* Added page variable and current page state also added isLoading state to hide page change
   const [currentPage, setCurrentPage] = useState<number>(1);
 
-  const { categories, loading } = useCategoryNav();
   const [data, setData] = useState<ReviewData[] | null>(null);
   const [rats, setRats] = useState<RatsData>();
   const [filterRating, setFilterRating] = useState<string>('all');
   const [filterView, setFilterView] = useState<string>('topReviews');
+  const [metaData, setMetaData] = useState<any>();
   const [filteredData, setFilteredData] = useState<ReviewData[] | null>(null);
   const [productName, setProductName] = useState<string>('');
   const [mountUI, setMountUI] = useState<boolean>(false);
@@ -72,15 +70,21 @@ const BuyersView = () => {
       }&pageSize=10`;
       fetch(url)
         .then((res) => res.json())
-        .then((data: ReviewApiResponse) => setData(data.data))
+        .then((data) => {
+          setMetaData(data.data.meta);
+          setData(data.data.items);
+        })
         .catch((e) => console.log(e));
     } else {
-      const url: string = `https://team-liquid-repo.onrender.com/api/review/shop/products/1/reviews/rating?rating=${filterRating}&pageNumber=${
+      const url: string = `https://team-liquid-repo.onrender.com/api/review/shop/products/${id}/reviews/rating?rating=${filterRating}&pageNumber=${
         currentPage - 1
       }&pageSize=10`;
       fetch(url)
         .then((res) => res.json())
-        .then((data: ReviewApiResponse) => setData(data.data))
+        .then((data) => {
+          setMetaData(data.data.meta);
+          setData(data.data.items);
+        })
         .catch((e) => console.log(e));
     }
   }, [mountUI, filterRating, filterView, currentPage, id]);
@@ -95,11 +99,11 @@ const BuyersView = () => {
   }, [id]);
 
   const ratingData = [
-    { rating: 5, users: rats?.fiveStar!, total: rats?.numberOfRating! },
-    { rating: 4, users: rats?.fourStar!, total: rats?.numberOfRating! },
-    { rating: 3, users: rats?.threeStar!, total: rats?.numberOfRating! },
-    { rating: 2, users: rats?.twoStar!, total: rats?.numberOfRating! },
-    { rating: 1, users: rats?.oneStar!, total: rats?.numberOfRating! },
+    { rating: 5, users: rats?.fiveStar! || 0, total: rats?.numberOfRating! },
+    { rating: 4, users: rats?.fourStar! || 0, total: rats?.numberOfRating! },
+    { rating: 3, users: rats?.threeStar! || 0, total: rats?.numberOfRating! },
+    { rating: 2, users: rats?.twoStar! || 0, total: rats?.numberOfRating! },
+    { rating: 1, users: rats?.oneStar! || 0, total: rats?.numberOfRating! },
   ];
 
   // ToDo: Remove all commented out code
@@ -151,15 +155,12 @@ const BuyersView = () => {
   return (
     <MainLayout activePage="Explore" showDashboardSidebar={false} showTopbar>
       <Container>
-        <div className="max-w-[1240px] hidden lg:block mx-auto my-0">
-          {/* from marketplace: this component you are using is from marketplace and it has been updated and we have updated it on your end also, this is important to allow sync without error take note  */}
-          <CategoriesNav navItems={categories} isLoading={loading} />
-        </div>
-        {!rats ? (
+        <NavDashBoard active="reviews" />
+        {!data ? (
           <div className=" h-[70vh] flex justify-center items-center">
             <Loader />
           </div>
-        ) : rats === null || rats.averageRating === undefined ? (
+        ) : data === null ? (
           <EmptyReviewPage />
         ) : (
           <div className="flex flex-col justify-center items-center md:mb-16">
@@ -170,25 +171,45 @@ const BuyersView = () => {
                   onClick={() => router.back()}
                 >
                   <Image src="/assets/reviews/return-icon.svg" width={32} height={32} alt="return" />
-                  {router.query.title}
+                  {/* Commented out title for testing */}
+                  {/* {router.query.title} */}
+                  Customer Reviews
                 </div>
               </div>
               <div className="flex flex-col md:flex-row lg:gap-24 md:gap-10 gap-4 mx-5">
                 <div className="flex flex-row md:flex-col gap-4 md:gap-8 lg:w-80 md:w-48">
                   <div>
-                    <RatingBar avgRating={rats?.averageRating!} verUser={100} />
+                    <RatingBar avgRating={rats?.averageRating! || 0} verUser={rats?.numberOfRating! || 0} />
+                    {/* <div className="md:hidden block">
+                      <p className="pt-6">Have any thoughts?</p>
+                      <Link
+                        href={`../create/${rats?.productId}`}
+                        className="flex text-sm md:text-base font-manropeB text-brand-green-pressed h-5 w-36 self-start"
+                      >
+                        <button className="hover:text-green-200">Write a Review!</button>
+                      </Link>
+                    </div> */}
                   </div>
                   <div className="flex flex-col gap-2">
                     {ratingData.map((data, index) => (
                       <RatingCard key={index} rating={data.rating} users={data.users} totalReviews={data.total} />
                     ))}
+                    {/* <div className="hidden md:block">
+                      <p className="pt-6">Have any thoughts?</p>
+                      <Link
+                        href={`../create/${rats?.productId}`}
+                        className="flexfont-manropeB text-brand-green-pressed h-5 w-36 self-start"
+                      >
+                        <button className="hover:text-green-200">Write a Review!</button>
+                      </Link>
+                    </div> */}
                   </div>
                 </div>
                 <div className="flex flex-col">
                   <div className="w-full justify-start">
                     <Filter
-                      rating={rats?.totalRating!}
-                      review={rats?.numberOfRating!}
+                      rating={rats?.totalRating! || 0}
+                      review={rats?.numberOfRating! || 0}
                       filterReview={(view, rating) => handleFilter(view, rating)}
                     />
                   </div>
@@ -196,7 +217,7 @@ const BuyersView = () => {
                     {!filteredData ? (
                       <Loader />
                     ) : filteredData?.length === 0 ? (
-                      <h2>No results</h2>
+                      <EmptyReviewPage />
                     ) : (
                       filteredData?.map((review) => (
                         <Review
@@ -216,17 +237,13 @@ const BuyersView = () => {
                 </div>
               </div>
             </div>
-            {data === null || data === undefined || data.length === 0 ? (
-              <div className=" w-0 h-0 m-0 p-0 hidden"></div>
-            ) : (
-              <Pagination
-                page={currentPage}
-                pages={data[0]?.numberOfPages}
-                activePage={currentPage}
-                visiblePaginatedBtn={3}
-                setPage={setPage}
-              />
-            )}
+            <Pagination
+              page={currentPage}
+              pages={metaData?.totalPages!}
+              activePage={currentPage}
+              visiblePaginatedBtn={3}
+              setPage={setPage}
+            />
           </div>
         )}
       </Container>
