+++ modules/dashboard/component/reviews/reviewPreviews/SellersPreview.tsx
@@ -1,7 +1,9 @@
 //* Removed some unnecessary imports
 
 import React, { useState, useEffect } from 'react';
+import NavDashBoard from '@modules/dashboard/component/Navbar';
 import Image from 'next/image';
+import Link from 'next/link';
 import { useRouter } from 'next/router';
 import RatingCard from '@modules/dashboard/component/reviews/review-page/RatingCard';
 import RatingBar from '@modules/dashboard/component/reviews/review-page/RatingBar';
@@ -13,7 +15,7 @@ import EmptyReviewPage from '@modules/dashboard/component/reviews/review-page/Em
 import Container from '@modules/auth/component/Container/Container';
 import Loader from '@ui/Loader';
 import { ReviewData, ReviewApiResponse, RatsData } from '../../../../../@types';
-import Breadcrumbs from '../../../../../components/Breadcrumbs';
+import { set } from 'nprogress';
 
 //* Moved type definitions to @types/index.d.ts
 const SellersPreview = () => {
@@ -34,6 +36,7 @@ const SellersPreview = () => {
 
   const [data, setData] = useState<ReviewData[] | null>(null);
   const [rats, setRats] = useState<RatsData>();
+  const [mountUI, setMountUI] = useState<boolean>(false);
 
   // ToDo: Remove all commented out code
   // const [total5Star, setTotal5Star] = useState<number>(0);
@@ -60,13 +63,15 @@ const SellersPreview = () => {
     if (id) {
       const url: string = `https://team-liquid-repo.onrender.com/api/review/shop/${id}/reviews?pageNumber=${
         currentPage - 1
-      }&pageSize=10`;
+      }&pageSize=1`;
       fetch(url)
         .then((res) => res.json())
-        .then((data: ReviewApiResponse) => setData(data.data))
+        .then((data) => {
+          setData(data.data.items);
+        })
         .catch((e) => console.log(e));
     }
-  }, [id]);
+  }, [mountUI, currentPage, id]);
   useEffect(() => {
     if (id) {
       const apiUrl: string = `https://team-liquid-repo.onrender.com/api/review/products/${id}/rating`;
@@ -78,20 +83,43 @@ const SellersPreview = () => {
   }, [id]);
 
   const ratingData = [
-    { rating: 5, users: rats?.fiveStar!, total: rats?.numberOfRating! },
-    { rating: 4, users: rats?.fourStar!, total: rats?.numberOfRating! },
-    { rating: 3, users: rats?.threeStar!, total: rats?.numberOfRating! },
-    { rating: 2, users: rats?.twoStar!, total: rats?.numberOfRating! },
-    { rating: 1, users: rats?.oneStar!, total: rats?.numberOfRating! },
+    { rating: 5, users: rats?.fiveStar! || 0, total: rats?.numberOfRating! },
+    { rating: 4, users: rats?.fourStar! || 0, total: rats?.numberOfRating! },
+    { rating: 3, users: rats?.threeStar! || 0, total: rats?.numberOfRating! },
+    { rating: 2, users: rats?.twoStar! || 0, total: rats?.numberOfRating! },
+    { rating: 1, users: rats?.oneStar! || 0, total: rats?.numberOfRating! },
   ];
+
+  // ToDo: Remove all commented out code
+  // useEffect(() => {
+  //   if (data) {
+  //     const total5 = data.filter((review) => review.rating === 5).length;
+  //     const total4 = data.filter((review) => review.rating === 4).length;
+  //     const total3 = data.filter((review) => review.rating === 3).length;
+  //     const total2 = data.filter((review) => review.rating === 2).length;
+  //     const total1 = data.filter((review) => review.rating === 1).length;
+  //     setTotal5Star(total5);
+  //     setTotal4Star(total4);
+  //     setTotal3Star(total3);
+  //     setTotal2Star(total2);
+  //     setTotal1Star(total1);
+  //   }
+  // }, [data]);
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setMountUI(true);
+    }, 1000);
+    return () => clearTimeout(timeout);
+  }, []);
 
   return (
     <div className="flex w-full justify-center items-center">
-      {!rats ? (
+      {!data ? (
         <div className="flex justify-center items-center">
           <Loader />
         </div>
-      ) : rats === null || rats.averageRating === undefined ? (
+      ) : data?.length === 0 ? (
         <EmptyReviewPage />
       ) : (
         <div className="w-full flex flex-col justify-center items-center md:mb-16">
@@ -99,36 +127,27 @@ const SellersPreview = () => {
             <div className="flex flex-col md:flex-row lg:gap-24 md:gap-10 gap-4 mx-5">
               <div className="flex flex-row md:flex-col gap-4 md:gap-8 lg:w-80 md:w-48">
                 <div>
-                  <RatingBar avgRating={rats?.averageRating!} verUser={100} />
-                </div>
-                <div className="flex flex-col gap-2">
-                  {ratingData.map((data, index) => (
-                    <RatingCard key={index} rating={data.rating} users={data.users} totalReviews={data.total} />
-                  ))}
+                  <RatingBar avgRating={rats?.averageRating!} verUser={rats?.numberOfRating || 0} />
                 </div>
               </div>
               <div className="w-full flex flex-col">
                 <div className="w-full mt-0 mx-1">
                   {!data ? (
                     <Loader />
-                  ) : data?.length === 0 ? (
-                    <h2>No results</h2>
                   ) : (
-                    data
-                      ?.slice(0, 1)
-                      .map((review) => (
-                        <SellerReview
-                          key={review.reviewId}
-                          buyerName={review.customerName}
-                          mainDate={review.createdAt}
-                          adminDate={review.reply?.createdAt}
-                          review={review.description}
-                          noOfStars={review.rating}
-                          shopReply={review.reply?.message}
-                          shopName={review.reply?.name}
-                          reviewId={review.reviewId}
-                        />
-                      ))
+                    data?.map((review) => (
+                      <SellerReview
+                        key={review.reviewId}
+                        buyerName={review.customerName}
+                        mainDate={review.createdAt}
+                        adminDate={review.reply?.createdAt}
+                        review={review.description}
+                        noOfStars={review.rating}
+                        shopReply={review.reply?.message}
+                        shopName={review.reply?.name}
+                        reviewId={review.reviewId}
+                      />
+                    ))
                   )}
                 </div>
               </div>
