+++ pages/view-components/review-components.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReviewDashboardCard from '@modules/dashboard/component/reviews/ReviewDashboardCard';
 import Filter from '@modules/dashboard/component/reviews/review-page/ReviewFilter';
-import Review from '@modules/dashboard/component/reviews/review-page/Review';
-import SellerReview from '@modules/dashboard/component/reviews/review-page/SellersReview';
+import BuyersPreview from '@modules/dashboard/component/reviews/reviewPreviews/BuyersPreview';
+import SellersPreview from '@modules/dashboard/component/reviews/reviewPreviews/SellersPreview';
 import ReviewForms from '@modules/dashboard/component/reviews/ReviewForm';
 import EmptyReviewPage from '@modules/dashboard/component/reviews/review-page/EmptyReviewPage';
 import RatingCard from '@modules/dashboard/component/reviews/review-page/RatingCard';
@@ -36,29 +36,11 @@ function reviewComponents() {
       </div>
       <div className="py-1 px-9 pb-10">
         <h3 className="text-3xl flex justify-center py-10">Buyers Review Component</h3>
-        <Review
-          reviewId="1"
-          buyerName="John Doe"
-          adminDate="2021-09-01"
-          mainDate="2021-09-01"
-          review="I recently purchased this digital product and I'm thrilled with it! It's been a game-changer for me. The user interface is incredibly intuitive, making it easy for even a tech novice like me to navigate. The product's functionality is top-notch, and it has exceeded my expectations. It's incredibly fast and efficient, saving me both time and frustration. Customer support has been exceptional."
-          noOfStars={3}
-        />
+        <BuyersPreview />
       </div>
       <div className="py-1 px-9 pb-10">
         <h3 className="text-3xl flex justify-center py-10">Sellers Review Component</h3>
-        <SellerReview
-          reviewId="1"
-          buyerName="John Doe"
-          adminDate="2021-09-01"
-          mainDate="2021-09-01"
-          review="I recently purchased this digital product and I'm thrilled with it! It's been a game-changer for me. The user interface is incredibly intuitive, making it easy for even a tech novice like me to navigate. The product's functionality is top-notch, and it has exceeded my expectations. It's incredibly fast and efficient, saving me both time and frustration. Customer support has been exceptional."
-          shopName={'Zuri Marketplace'}
-          shopReply={
-            'Having this product is the best thing that has happened to me in a very long time. Thank you so much for this product. The shipping and delivery was also very good. But there a few tweaks that this can actually have though. Having this product is the best thing that has happened to me in a very long time. Thank you so much for this product.'
-          }
-          noOfStars={3}
-        />
+        <SellersPreview />
       </div>
       <div className="py-1 px-9 pb-10">
         <h3 className="text-3xl flex justify-center py-10">Review Form</h3>
